post feedback commit
split classes into multiple files, using objects for the transaction history, returning string rather than print to console
